backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# LangChain + LangGraph
langchain==0.3.1
langchain-community==0.3.1
langchain-ollama==0.2.0
langgraph==0.2.28
langchain-chroma==0.1.4

# Vector store
chromadb==0.5.11

# File parsing
pypdf==5.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
unstructured==0.15.13

# MySQL
mysql-connector-python==9.0.0
sqlalchemy==2.0.35

# Async utilities
aiofiles==24.1.0
httpx==0.27.2

# Embeddings (local via Ollama)
sentence-transformers==3.1.1
